+++ parrot/alembic/versions/fd7c085ab081_drop_message_timestamp_column.py
@@ -36,7 +36,7 @@ def downgrade() -> None:
 		"message",
 		sa.Column(
 			"timestamp",
-			sa.Integer(),
+			sa.BigInteger(),
 			nullable=False,
 			server_default="0",
 		),
